Fix database setup and write basic search and get

+++ src/main/java/itmo/airbnb/repos/CalendarRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.Calendar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
+public interface CalendarRepository extends JpaRepository<Calendar, Long> {
 }
