+++ src/main/java/itmo/airbnb/repos/GuestRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface GuestRepository extends JpaRepository<Guest, Integer> {
+public interface GuestRepository extends JpaRepository<Guest, Long> {
 }
