+++ src/main/java/itmo/airbnb/repos/HostInfoRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.HostInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface HostInfoRepository extends JpaRepository<HostInfo, Integer> {
+public interface HostInfoRepository extends JpaRepository<HostInfo, Long> {
 }
