+++ src/main/java/itmo/airbnb/repos/HostRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.Host;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface HostRepository extends JpaRepository<Host, Integer> {
+public interface HostRepository extends JpaRepository<Host, Long> {
 }
