+++ src/main/java/itmo/airbnb/repos/ListingGuestRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.ListingGuest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ListingGuestRepository extends JpaRepository<ListingGuest, Integer> {
+public interface ListingGuestRepository extends JpaRepository<ListingGuest, Long> {
 }
