+++ src/main/java/itmo/airbnb/repos/ListingPriceRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.ListingPrice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ListingPriceRepository extends JpaRepository<ListingPrice, Integer> {
+public interface ListingPriceRepository extends JpaRepository<ListingPrice, Long> {
 }
