+++ src/main/java/itmo/airbnb/repos/UserLoginDataRepository.java
@@ -4,5 +4,5 @@ import itmo.airbnb.domain.UserLoginData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface UserLoginDataRepository extends JpaRepository<UserLoginData, Integer> {
+public interface UserLoginDataRepository extends JpaRepository<UserLoginData, Long> {
 }
